requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==39.0.1
cssselect==1.2.0
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.3.0
Django==4.1.6
django-allauth==0.52.0
django-bootstrap-v5==1.0.11
django-cors-headers==3.14.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.8.0
idna==3.4
jieba3k==0.35.1
joblib==1.2.0
lxml==4.9.2
newsapi-python==0.2.7
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.23.3
oauthlib==3.2.2
pandas==2.0.0
pbr==5.10.0
Pillow==9.2.0
platformdirs==2.5.2
postgres==4.0
psycopg2==2.9.5
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pycparser==2.21
pygame==2.1.2
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0
regex==2023.3.23
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
sgmllib3k==1.0.0
six==1.16.0
soupsieve==2.4
sqlparse==0.4.3
stevedore==4.0.0
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==3.4.0
torchvision==0.13.1
tqdm==4.65.0
typing_extensions==4.4.0
tzdata==2023.3
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
